+++ cmd/kioskmate/instance_lock_linux.go
@@ -1,0 +1,31 @@
+//go:build linux
+
+package main
+
+import (
+	"fmt"
+	"io"
+	"os"
+	"path/filepath"
+
+	"golang.org/x/sys/unix"
+)
+
+func acquireInstanceLock(path string) (io.Closer, error) {
+	if err := os.MkdirAll(filepath.Dir(path), 0o700); err != nil {
+		return nil, err
+	}
+	file, err := os.OpenFile(path, os.O_CREATE|os.O_RDWR, 0o600)
+	if err != nil {
+		return nil, err
+	}
+	if err := unix.Flock(int(file.Fd()), unix.LOCK_EX|unix.LOCK_NB); err != nil {
+		_ = file.Close()
+		return nil, fmt.Errorf("lock %s: %w", path, err)
+	}
+	if err := file.Truncate(0); err == nil {
+		_, _ = fmt.Fprintf(file, "%d\n", os.Getpid())
+		_ = file.Sync()
+	}
+	return file, nil
+}